+++ api/src/main/java/com/example/myecommerce/api/dto/Comment/CommentReqDto.java
@@ -1,8 +1,8 @@
 package com.example.myecommerce.api.dto.Comment;
 
-import com.example.myecoomerce.myecommercecore.Domain.Comment.Comment;
-import com.example.myecoomerce.myecommercecore.Domain.Product.Product;
-import com.example.myecoomerce.myecommercecore.Domain.User.User;
+import com.example.myecoomerce.myecommercecore.Comment.Comment;
+import com.example.myecoomerce.myecommercecore.Product.Product;
+import com.example.myecoomerce.myecommercecore.User.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
