+++ api/src/main/java/com/example/myecommerce/api/dto/Product/ProductListByCategoryResDto.java
@@ -1,6 +1,6 @@
 package com.example.myecommerce.api.dto.Product;
 
-import com.example.myecoomerce.myecommercecore.Domain.Product.Product;
+import com.example.myecoomerce.myecommercecore.Product.Product;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
