+++ api/src/main/java/com/example/myecommerce/api/service/User/CustomUserDetailService.java
@@ -1,6 +1,6 @@
 package com.example.myecommerce.api.service.User;
 
-import com.example.myecoomerce.myecommercecore.Domain.User.UserRepository;
+import com.example.myecoomerce.myecommercecore.User.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
