+++ api/src/main/java/com/example/myecommerce/api/service/User/UserService.java
@@ -3,9 +3,9 @@ package com.example.myecommerce.api.service.User;
 import com.example.myecommerce.api.dto.User.PasswordUpdateReqDto;
 import com.example.myecommerce.api.dto.User.UserRegisterReqDto;
 import com.example.myecommerce.api.dto.User.UserUpdateReqDto;
-import com.example.myecoomerce.myecommercecore.Domain.User.Role;
-import com.example.myecoomerce.myecommercecore.Domain.User.User;
-import com.example.myecoomerce.myecommercecore.Domain.User.UserRepository;
+import com.example.myecoomerce.myecommercecore.User.Role;
+import com.example.myecoomerce.myecommercecore.User.User;
+import com.example.myecoomerce.myecommercecore.User.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.core.context.SecurityContextHolder;
