+++ api/src/main/resources-local/application.properties
@@ -1,5 +1,5 @@
 profile=local
-#spring.profiles.include=core
+spring.profiles.include=core
 spring.profiles.active=local
 
 spring.datasource.url=jdbc:mysql://localhost:3306/Ecommerce?serverTimezone=UTC&characterEncoding=UTF-8&useSSL=false&allowPublicKeyRetrieval=true
