+++ api/src/test/java/UserApiControllerTest.java
@@ -2,9 +2,9 @@ import com.example.myecommerce.api.config.Security.JwtTokenProvider;
 import com.example.myecommerce.api.dto.User.LoginReqDto;
 import com.example.myecommerce.api.dto.User.UserRegisterReqDto;
 import com.example.myecommerce.api.dto.User.UserResDto;
-import com.example.myecoomerce.myecommercecore.Domain.User.Role;
-import com.example.myecoomerce.myecommercecore.Domain.User.User;
-import com.example.myecoomerce.myecommercecore.Domain.User.UserRepository;
+import com.example.myecoomerce.myecommercecore.User.Role;
+import com.example.myecoomerce.myecommercecore.User.User;
+import com.example.myecoomerce.myecommercecore.User.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.After;
 import org.junit.Before;
