+++ build.gradle
@@ -69,11 +69,21 @@ project(':api') {
 project(':view') {
     // 공통 코드
 
+    bootJar { enabled = true }
+    jar { enabled = false }
+
+    dependencies {
+        implementation project(':ecommerce-core')
+    }
+}
+
+project(':common') {
+    // 공통 코드
+
     bootJar { enabled = false } // core 은 bootJar 로 패키징 할 필요 없음
     jar { enabled = true }
 
     dependencies {
         implementation project(':ecommerce-core')
-        implementation project(':api')
     }
 }
