+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/ImageFile/ImageFile.java
@@ -1,7 +1,7 @@
-package com.example.myecoomerce.myecommercecore.Domain.ImageFile;
+package com.example.myecoomerce.myecommercecore.ImageFile;
 
-import com.example.myecoomerce.myecommercecore.Domain.BaseTimeEntity;
-import com.example.myecoomerce.myecommercecore.Domain.Product.Product;
+import com.example.myecoomerce.myecommercecore.BaseTimeEntity;
+import com.example.myecoomerce.myecommercecore.Product.Product;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Builder;
 import lombok.Getter;
