+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Order/OrderRepository.java
@@ -1,4 +1,4 @@
-package com.example.myecoomerce.myecommercecore.Domain.Order;
+package com.example.myecoomerce.myecommercecore.Order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
