+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Product/Product.java
@@ -1,11 +1,11 @@
-package com.example.myecoomerce.myecommercecore.Domain.Product;
+package com.example.myecoomerce.myecommercecore.Product;
 
-import com.example.myecoomerce.myecommercecore.Domain.BaseTimeEntity;
-import com.example.myecoomerce.myecommercecore.Domain.Category.Category;
-import com.example.myecoomerce.myecommercecore.Domain.Comment.Comment;
-import com.example.myecoomerce.myecommercecore.Domain.ImageFile.ImageFile;
-import com.example.myecoomerce.myecommercecore.Domain.Order.Order;
-import com.example.myecoomerce.myecommercecore.Domain.User.User;
+import com.example.myecoomerce.myecommercecore.BaseTimeEntity;
+import com.example.myecoomerce.myecommercecore.Category.Category;
+import com.example.myecoomerce.myecommercecore.Comment.Comment;
+import com.example.myecoomerce.myecommercecore.ImageFile.ImageFile;
+import com.example.myecoomerce.myecommercecore.Order.Order;
+import com.example.myecoomerce.myecommercecore.User.User;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.Builder;
