+++ view/src/main/java/com/example/myecommerce/view/contoller/AdminController.java
@@ -1,30 +1,25 @@
 package com.example.myecommerce.view.contoller;
 
-import com.example.myecommerce.api.dto.Product.ProductListResDto;
-import com.example.myecommerce.api.service.Product.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
-import java.util.List;
 
 @RequestMapping("/admin")
 @Controller
 @RequiredArgsConstructor
 public class AdminController {
 
-    private final ProductService productService;
+//    private final ProductService productService;
 
-    @RequestMapping("/")
-    public ModelAndView index(ModelAndView mv) {
-        List<ProductListResDto> list = productService.findAll();
-        mv.setViewName("board");
-        mv.addObject("productList", list);
-        mv.addObject("title", "상품 관리");
-        return mv;
-    }
+//    @RequestMapping("/")
+//    public ModelAndView index(ModelAndView mv) {
+//        List<ProductListResDto> list = productService.findAll();
+//        mv.setViewName("board");
+//        mv.addObject("productList", list);
+//        mv.addObject("title", "상품 관리");
+//        return mv;
+//    }
 
     @RequestMapping("/hello")
     public String hello() {
