--- view/src/main/resources/application-core.properties
@@ -1,14 +1,0 @@
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
-spring.jpa.open-in-view=false
-spring.thymeleaf.cache=false
-spring.datasource.driverClassName=com.mysql.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=none
-logging.level.org.hibernate.sql=info
-
-file.upload-dir=/Users/gim-yuwan/Documents/project/my-ecommerce/src/main/resources/static/file
-spring.h2.console.enabled=true
-spring.jwt.secret=MySecret$12345
-spring.servlet.multipart.max-file-size=200MB
-spring.servlet.multipart.max-request-size=215MB
